Add solution for problem `B - Count Subrectangles` Link: https://codeforces.com/contest/1323/problem/B Tags: binary search, greedy, implementation, *1500 Ref: https://codeforces.com/contest/1323/submission/74654593

+++ codeforces/1323/B.cpp
@@ -14,8 +14,8 @@ int main()
     for(int &i : arrB)
         cin >> i;
     
-    map<int, ll> onesA;
-    map<int, ll> onesB;
+    map<int, int> onesA;
+    map<int, int> onesB;
     int cur = 0;
     for(int i = 0; i < n; i++) {
         if(arrA[i] == 1)
